requirements.txt
## The following requirements were added by pip freeze:
astroid==1.5.3
behave==1.2.5
behave-django==1.0.0
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.6
django-parler==1.8
djangorestframework==3.7.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
parse==1.8.2
parse-type==0.4.2
pylint==1.7.4
pytz==2017.3
six==1.11.0
wrapt==1.10.11
